backend/requirements.txt
fastapi
uvicorn
langchain
langchain-aws
langgraph
python-multipart
boto3
requests
psycopg2-binary
pydantic
pytest
pytest-asyncio
pytest-cov